refactor(patina): keep Nuxt preset scoping within source length limits
