+++ crates/vize/tests/lint_config_cli.rs
@@ -328,59 +328,3 @@ await nextTick()
 
     let _ = fs::remove_dir_all(project_root);
 }
-
-#[test]
-fn lint_limits_nuxt_link_guidance_to_nuxt_preset() {
-    let project_root = temp_project_dir("nuxt-link-preset-boundary");
-    write_project_file(
-        &project_root,
-        "app/pages/index.vue",
-        r#"<template><a href="/settings">Settings</a></template>"#,
-    );
-    let ecosystem = Command::new(env!("CARGO_BIN_EXE_vize"))
-        .current_dir(&project_root)
-        .args([
-            "lint",
-            "--preset",
-            "ecosystem",
-            "--no-config",
-            "--format",
-            "json",
-            "app/pages/index.vue",
-        ])
-        .output()
-        .unwrap();
-    assert!(ecosystem.status.success(), "{}", output_details(&ecosystem));
-    assert!(
-        !String::from_utf8_lossy(&ecosystem.stdout).contains("ecosystem/nuxt-prefer-nuxt-link"),
-        "{}",
-        output_details(&ecosystem)
-    );
-
-    write_project_file(
-        &project_root,
-        "nuxt.config.ts",
-        "export default defineNuxtConfig({})\n",
-    );
-    let nuxt = Command::new(env!("CARGO_BIN_EXE_vize"))
-        .current_dir(&project_root)
-        .args([
-            "lint",
-            "--preset",
-            "nuxt",
-            "--no-config",
-            "--format",
-            "json",
-            "app/pages/index.vue",
-        ])
-        .output()
-        .unwrap();
-    assert!(nuxt.status.success(), "{}", output_details(&nuxt));
-    assert!(
-        String::from_utf8_lossy(&nuxt.stdout).contains("ecosystem/nuxt-prefer-nuxt-link"),
-        "{}",
-        output_details(&nuxt)
-    );
-
-    let _ = fs::remove_dir_all(project_root);
-}
