+++ crates/vize_patina/src/linter/tests.rs
@@ -11,6 +11,7 @@ mod no_mutating_props;
 mod no_top_level_ref;
 mod nuxt;
 mod nuxt_config_order;
+mod nuxt_prefer_nuxt_link;
 mod recovered_parse_analysis;
 mod script;
 mod severity_overrides;
