+++ crates/vize_patina/src/rule.rs
@@ -414,8 +414,7 @@ impl RuleRegistry {
     /// Create registry with all available rules (including opt-in).
     pub fn with_all() -> Self {
         let mut registry = Self::with_opinionated();
-        crate::rules::ecosystem::register(&mut registry);
-        crate::rules::ecosystem::register_nuxt(&mut registry);
+        crate::rules::ecosystem::register_all(&mut registry);
 
         registry
     }
@@ -423,8 +422,7 @@ impl RuleRegistry {
     /// Create registry with Nuxt-friendly rules (auto-imports enabled).
     pub fn with_nuxt() -> Self {
         let mut registry = Self::with_happy_path();
-        crate::rules::opinionated::register_nuxt(&mut registry);
-        crate::rules::ecosystem::register_nuxt(&mut registry);
+        crate::rules::register_nuxt(&mut registry);
 
         registry
     }
